Update to lvgl 9.0.0
SDL builds and runs fine on Linux.
X11 builds and runs fine on Linux except for erroring on close.
Windows native builds fine on Linux and then runs correctly on Windows including working keyboard and correct speed unlike with 8.x.
WIndows native and SDL builds on MSYS2 fail because Make cannot handle the final build command as it is too long, this is a common issue on Windows and so solutions definitely exist.